Move custom jsoniter code into json_codec.go.
